Split widgets page into one subpage per widget
