+++ ui/src/assets/widgets/editor.scss
@@ -23,4 +23,8 @@
       font-size: 13px;
     }
   }
+
+  &--fill-height {
+    height: 100%;
+  }
 }
