+++ ui/src/plugins/dev.perfetto.WidgetsPage/demos/hotkey_demo.ts
@@ -1,0 +1,47 @@
+// Copyright (C) 2025 The Android Open Source Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import m from 'mithril';
+import {HotkeyGlyphs} from '../../../widgets/hotkey_glyphs';
+import {Hotkey} from '../../../base/hotkeys';
+import {EnumOption, renderWidgetShowcase} from '../widgets_page_utils';
+
+export function renderHotkey(): m.Children {
+  return [
+    m(
+      '.pf-widget-intro',
+      m('h1', 'Hotkey'),
+      m(
+        'p',
+        'Display keyboard shortcuts in a platform-appropriate format (e.g., Cmd on Mac, Ctrl on Windows).',
+      ),
+    ),
+    renderWidgetShowcase({
+      renderWidget: ({platform, hotkey}) => {
+        if (platform === 'auto') {
+          return m(HotkeyGlyphs, {hotkey: hotkey as Hotkey});
+        } else {
+          return m(HotkeyGlyphs, {
+            hotkey: hotkey as Hotkey,
+            spoof: platform,
+          });
+        }
+      },
+      initialOpts: {
+        hotkey: 'Mod+Shift+P',
+        platform: new EnumOption('auto', ['auto', 'Mac', 'PC'] as const),
+      },
+    }),
+  ];
+}
