+++ ui/src/plugins/dev.perfetto.WidgetsPage/index.ts
@@ -23,7 +23,7 @@ export default class implements PerfettoPlugin {
   static onActivate(app: App): void {
     app.pages.registerPage({
       route: '/widgets',
-      render: () => m(WidgetsPage, {app}),
+      render: (subpage) => m(WidgetsPage, {app, subpage}),
     });
     app.sidebar.addMenuItem({
       section: 'navigation',
