+++ ui/src/plugins/dev.perfetto.WidgetsPage/styles.scss
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 .pf-widgets-page {
-  padding: 20px;
-  font-size: 16px;
-  overflow: auto;
+  display: flex;
+  height: 100%;
+  overflow: hidden;
 
   h1 {
-    margin: 32px 0 0 0;
+    margin: 24px 0 0 0;
     font-size: 28px;
   }
 
@@ -27,50 +27,104 @@
     font-size: 24px;
   }
 
+  h3 {
+    margin: 12px 0 0 0;
+    font-size: 20px;
+  }
+
+  p,
   ul {
-    margin-block-start: 5px;
-    margin-block-end: 0px;
-    padding-inline-start: 0px;
+    margin: 8px 0 0 0;
   }
 
   li {
-    list-style-type: none;
-    margin: 2px 0 0 0;
+    margin-bottom: 4px;
+  }
+
+  strong {
+    font-weight: 400;
+  }
+
+  &__option {
+    padding-inline: 16px;
+    margin-bottom: 4px;
+  }
+
+  &__options-title {
+    padding-inline: 16px;
+    margin: 8px 0;
+    font-size: 18px;
+    font-weight: 500;
   }
 
   .pf-widget-block {
+    margin-top: 16px;
     display: flex;
     flex-direction: row;
+    gap: 10px;
   }
 
   .pf-widget-controls {
-    margin: 10px;
+    flex: 0;
+    min-width: 250px;
+    overflow: hidden;
+
+    background: var(--pf-color-background-secondary);
+    border-radius: 3px;
+    padding-block: 8px;
   }
 
   .pf-widget-container {
-    display: flex;
+    min-height: 300px;
+    max-height: 500px;
 
-    // Put the widget in the center of the container
-    align-items: center;
-    justify-content: center;
-
-    min-width: 300px;
-    max-width: 800px;
-    min-height: 250px;
-    max-height: 1000px;
     border-radius: 3px;
-    box-shadow: inset 2px 2px 10px #00000020;
-    border: dashed 1px gray;
-    margin: 10px 0 10px 0;
-    padding: 16px;
+    box-shadow: inset 2px 2px 10px var(--pf-color-box-shadow);
+    border: dashed 1px var(--pf-color-border);
+    padding: 24px;
     overflow: hidden;
-  }
 
-  .pf-widget-container--wide {
-    min-width: 600px;
+    flex: 1;
   }
 
   .pf-virtual-canvas {
     height: 100px;
+  }
+
+  &__menu {
+    width: 200px;
+    border-right: 1px solid var(--pf-color-border);
+    overflow-y: auto;
+    padding: 8px;
+    flex-shrink: 0;
+  }
+
+  &__menu-item {
+    padding: 4px 8px;
+    cursor: pointer;
+    border-radius: 4px;
+    margin-bottom: 2px;
+    background-color: transparent;
+
+    &:hover {
+      background-color: color_hover(transparent);
+    }
+
+    &.pf-active {
+      background-color: color_active(transparent);
+      font-weight: 500;
+    }
+  }
+
+  &__content-container {
+    flex: 1;
+    overflow-y: auto;
+  }
+
+  &__content {
+    flex: 1;
+    padding-inline: 16px;
+    max-width: 900px;
+    padding-bottom: 32px;
   }
 }
