+++ ui/src/plugins/dev.perfetto.WidgetsPage/widgets_page.ts
@@ -13,1835 +13,153 @@
 // limitations under the License.
 
 import m from 'mithril';
-import {Hotkey, Platform} from '../../base/hotkeys';
-import {Icons} from '../../base/semantic_icons';
-import {Anchor} from '../../widgets/anchor';
-import {
-  Button,
-  ButtonAttrs,
-  ButtonBar,
-  ButtonGroup,
-  ButtonVariant,
-} from '../../widgets/button';
-import {Callout} from '../../widgets/callout';
-import {Checkbox} from '../../widgets/checkbox';
-import {Editor} from '../../widgets/editor';
-import {EmptyState} from '../../widgets/empty_state';
-import {Form, FormLabel} from '../../widgets/form';
-import {HotkeyGlyphs} from '../../widgets/hotkey_glyphs';
-import {Icon} from '../../widgets/icon';
-import {Menu, MenuDivider, MenuItem, PopupMenu} from '../../widgets/menu';
-import {showModal} from '../../widgets/modal';
-import {
-  MultiSelect,
-  MultiSelectDiff,
-  PopupMultiSelect,
-} from '../../widgets/multiselect';
-import {Popup, PopupPosition} from '../../widgets/popup';
-import {Portal} from '../../widgets/portal';
-import {Select} from '../../widgets/select';
-import {Spinner} from '../../widgets/spinner';
-import {Switch} from '../../widgets/switch';
-import {TextInput} from '../../widgets/text_input';
-import {MultiParagraphText, TextParagraph} from '../../widgets/text_paragraph';
-import {LazyTreeNode, Tree, TreeNode} from '../../widgets/tree';
-import {VegaView} from '../../components/widgets/vega_view';
-import {TreeTable, TreeTableAttrs} from '../../components/widgets/treetable';
-import {Intent} from '../../widgets/common';
-import {TagInput} from '../../widgets/tag_input';
-import {SegmentedButtons} from '../../widgets/segmented_buttons';
-import {MiddleEllipsis} from '../../widgets/middle_ellipsis';
-import {Chip} from '../../widgets/chip';
-import {TrackShell} from '../../widgets/track_shell';
-import {CopyableLink} from '../../widgets/copyable_link';
-import {CopyToClipboardButton} from '../../widgets/copy_to_clipboard_button';
-import {VirtualOverlayCanvas} from '../../widgets/virtual_overlay_canvas';
-import {SplitPanel, Tab} from '../../widgets/split_panel';
-import {parseAndPrintTree} from '../../base/perfetto_sql_lang/language';
-import {CursorTooltip} from '../../widgets/cursor_tooltip';
-import {MultiselectInput} from '../../widgets/multiselect_input';
-import {
-  DataGrid,
-  DataGridAttrs,
-} from '../../components/widgets/data_grid/data_grid';
-import {SQLDataSource} from '../../components/widgets/data_grid/sql_data_source';
+import {classNames} from '../../base/classnames';
 import {App} from '../../public/app';
-import {Engine} from '../../trace_processor/engine';
-import {Card, CardStack} from '../../widgets/card';
-import {Stack} from '../../widgets/stack';
-import {Tooltip} from '../../widgets/tooltip';
-import {TabStrip} from '../../widgets/tabs';
-import {CodeSnippet} from '../../widgets/code_snippet';
-import {
-  EnumOption,
-  renderWidgetShowcase,
-  WidgetShowcase,
-} from './widget_page_utils';
-import {renderGridDemo} from './grid_demo';
-import {renderDataGridDemo} from './data_grid_demo';
-import {NodeGraphDemo} from './nodegraph_demo';
+import {anchor} from './demos/anchor_demo';
+import {renderButtonDemo} from './demos/button_demo';
+import {renderButtonGroupDemo} from './demos/button_group_demo';
+import {renderCallout} from './demos/callout_demo';
+import {renderCard} from './demos/card_demo';
+import {cardStack} from './demos/cardstack_demo';
+import {renderCheckbox} from './demos/checkbox_demo';
+import {renderChip} from './demos/chip_demo';
+import {renderCodeSnippet} from './demos/code_snippet_demo';
+import {renderCopyableLink} from './demos/copyable_link_demo';
+import {cursorTooltip} from './demos/cursor_tooltip_demo';
+import {renderDataGrid} from './demos/datagrid_demo';
+import {renderEditor} from './demos/editor_demo';
+import {renderEmptyState} from './demos/empty_state_demo';
+import {renderForm} from './demos/form_demo';
+import {renderGrid} from './demos/grid_demo';
+import {renderHotkey} from './demos/hotkey_demo';
+import {renderIcon} from './demos/icon_demo';
+import {renderMenu} from './demos/menu_demo';
+import {renderMiddleEllipsis} from './demos/middle_ellipsis_demo';
+import {renderModal} from './demos/modal_demo';
+import {renderMultiselect} from './demos/multiselect_demo';
+import {renderNodeGraph} from './demos/nodegraph_demo';
+import {renderPopup} from './demos/popup_demo';
+import {popupMenuDemo} from './demos/popup_menu_demo';
+import {renderPortal} from './demos/portal_demo';
+import {segmentedButtons} from './demos/segmented_buttons_demo';
+import {renderSelect} from './demos/select_demo';
+import {renderSpinner} from './demos/spinner_demo';
+import {renderSplitPanel} from './demos/split_panel_demo';
+import {renderSwitch} from './demos/switch_demo';
+import {renderTabStrip} from './demos/tabstrip_demo';
+import {renderTagInput} from './demos/tag_input_demo';
+import {renderTextInput} from './demos/text_input_demo';
+import {renderTextParagraph} from './demos/text_paragraph_demo';
+import {renderTooltip} from './demos/tooltip_demo';
+import {renderTrackShell} from './demos/track_shell_demo';
+import {renderTree} from './demos/tree_demo';
+import {renderTreeTable} from './demos/treetable_demo';
+import {renderVegaView} from './demos/vega_view_demo';
+import {renderVirtualCanvas} from './demos/virtual_canvas_demo';
 
-const DATA_ENGLISH_LETTER_FREQUENCY = {
-  table: [
-    {category: 'a', amount: 8.167},
-    {category: 'b', amount: 1.492},
-    {category: 'c', amount: 2.782},
-    {category: 'd', amount: 4.253},
-    {category: 'e', amount: 12.7},
-    {category: 'f', amount: 2.228},
-    {category: 'g', amount: 2.015},
-    {category: 'h', amount: 6.094},
-    {category: 'i', amount: 6.966},
-    {category: 'j', amount: 0.253},
-    {category: 'k', amount: 1.772},
-    {category: 'l', amount: 4.025},
-    {category: 'm', amount: 2.406},
-    {category: 'n', amount: 6.749},
-    {category: 'o', amount: 7.507},
-    {category: 'p', amount: 1.929},
-    {category: 'q', amount: 0.095},
-    {category: 'r', amount: 5.987},
-    {category: 's', amount: 6.327},
-    {category: 't', amount: 9.056},
-    {category: 'u', amount: 2.758},
-    {category: 'v', amount: 0.978},
-    {category: 'w', amount: 2.36},
-    {category: 'x', amount: 0.25},
-    {category: 'y', amount: 1.974},
-    {category: 'z', amount: 0.074},
-  ],
-};
-
-const DATA_POLISH_LETTER_FREQUENCY = {
-  table: [
-    {category: 'a', amount: 8.965},
-    {category: 'b', amount: 1.482},
-    {category: 'c', amount: 3.988},
-    {category: 'd', amount: 3.293},
-    {category: 'e', amount: 7.921},
-    {category: 'f', amount: 0.312},
-    {category: 'g', amount: 1.377},
-    {category: 'h', amount: 1.072},
-    {category: 'i', amount: 8.286},
-    {category: 'j', amount: 2.343},
-    {category: 'k', amount: 3.411},
-    {category: 'l', amount: 2.136},
-    {category: 'm', amount: 2.911},
-    {category: 'n', amount: 5.6},
-    {category: 'o', amount: 7.59},
-    {category: 'p', amount: 3.101},
-    {category: 'q', amount: 0.003},
-    {category: 'r', amount: 4.571},
-    {category: 's', amount: 4.263},
-    {category: 't', amount: 3.966},
-    {category: 'u', amount: 2.347},
-    {category: 'v', amount: 0.034},
-    {category: 'w', amount: 4.549},
-    {category: 'x', amount: 0.019},
-    {category: 'y', amount: 3.857},
-    {category: 'z', amount: 5.62},
-  ],
-};
-
-const DATA_EMPTY = {};
-
-const SPEC_BAR_CHART = `
-{
-  "$schema": "https://vega.github.io/schema/vega/v5.json",
-  "description": "A basic bar chart example, with value labels shown upon mouse hover.",
-  "width": 400,
-  "height": 200,
-  "padding": 5,
-
-  "data": [
-    {
-      "name": "table"
-    }
-  ],
-
-  "signals": [
-    {
-      "name": "tooltip",
-      "value": {},
-      "on": [
-        {"events": "rect:mouseover", "update": "datum"},
-        {"events": "rect:mouseout",  "update": "{}"}
-      ]
-    }
-  ],
-
-  "scales": [
-    {
-      "name": "xscale",
-      "type": "band",
-      "domain": {"data": "table", "field": "category"},
-      "range": "width",
-      "padding": 0.05,
-      "round": true
-    },
-    {
-      "name": "yscale",
-      "domain": {"data": "table", "field": "amount"},
-      "nice": true,
-      "range": "height"
-    }
-  ],
-
-  "axes": [
-    { "orient": "bottom", "scale": "xscale" },
-    { "orient": "left", "scale": "yscale" }
-  ],
-
-  "marks": [
-    {
-      "type": "rect",
-      "from": {"data":"table"},
-      "encode": {
-        "enter": {
-          "x": {"scale": "xscale", "field": "category"},
-          "width": {"scale": "xscale", "band": 1},
-          "y": {"scale": "yscale", "field": "amount"},
-          "y2": {"scale": "yscale", "value": 0}
-        },
-        "update": {
-          "fill": {"value": "steelblue"}
-        },
-        "hover": {
-          "fill": {"value": "red"}
-        }
-      }
-    },
-    {
-      "type": "text",
-      "encode": {
-        "enter": {
-          "align": {"value": "center"},
-          "baseline": {"value": "bottom"},
-          "fill": {"value": "#333"}
-        },
-        "update": {
-          "x": {"scale": "xscale", "signal": "tooltip.category", "band": 0.5},
-          "y": {"scale": "yscale", "signal": "tooltip.amount", "offset": -2},
-          "text": {"signal": "tooltip.amount"},
-          "fillOpacity": [
-            {"test": "datum === tooltip", "value": 0},
-            {"value": 1}
-          ]
-        }
-      }
-    }
-  ]
-}
-`;
-
-const SPEC_BAR_CHART_LITE = `
-{
-  "$schema": "https://vega.github.io/schema/vega-lite/v5.json",
-  "description": "A simple bar chart with embedded data.",
-  "data": {
-    "name": "table"
-  },
-  "mark": "bar",
-  "encoding": {
-    "x": {"field": "category", "type": "nominal", "axis": {"labelAngle": 0}},
-    "y": {"field": "amount", "type": "quantitative"}
-  }
-}
-`;
-
-const SPEC_BROKEN = `{
-  "description": 123
-}
-`;
-
-enum SpecExample {
-  BarChart = 'Barchart',
-  BarChartLite = 'Barchart (Lite)',
-  Broken = 'Broken',
+interface WidgetSection {
+  readonly id: string;
+  readonly label: string;
+  readonly view: (app: App) => m.Children;
 }
 
-enum DataExample {
-  English = 'English',
-  Polish = 'Polish',
-  Empty = 'Empty',
-}
-
-function arg<T>(
-  anyArg: unknown,
-  valueIfTrue: T,
-  valueIfFalse: T | undefined = undefined,
-): T | undefined {
-  return Boolean(anyArg) ? valueIfTrue : valueIfFalse;
-}
-
-function getExampleSpec(example: SpecExample): string {
-  switch (example) {
-    case SpecExample.BarChart:
-      return SPEC_BAR_CHART;
-    case SpecExample.BarChartLite:
-      return SPEC_BAR_CHART_LITE;
-    case SpecExample.Broken:
-      return SPEC_BROKEN;
-    default:
-      const exhaustiveCheck: never = example;
-      throw new Error(`Unhandled case: ${exhaustiveCheck}`);
-  }
-}
-
-function getExampleData(example: DataExample) {
-  switch (example) {
-    case DataExample.English:
-      return DATA_ENGLISH_LETTER_FREQUENCY;
-    case DataExample.Polish:
-      return DATA_POLISH_LETTER_FREQUENCY;
-    case DataExample.Empty:
-      return DATA_EMPTY;
-    default:
-      const exhaustiveCheck: never = example;
-      throw new Error(`Unhandled case: ${exhaustiveCheck}`);
-  }
-}
-
-const options: {[key: string]: boolean} = {
-  foobar: false,
-  foo: false,
-  bar: false,
-  baz: false,
-  qux: false,
-  quux: false,
-  corge: false,
-  grault: false,
-  garply: false,
-  waldo: false,
-  fred: false,
-  plugh: false,
-  xyzzy: false,
-  thud: false,
-};
-
-let currentTab: string = 'foo';
-
-function PortalButton() {
-  let portalOpen = false;
-
-  return {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    view: function ({attrs}: any) {
-      const {zIndex = true, absolute = true, top = true} = attrs;
-      return [
-        m(Button, {
-          label: 'Toggle Portal',
-          onclick: () => {
-            portalOpen = !portalOpen;
-          },
-        }),
-        portalOpen &&
-          m(
-            Portal,
-            {
-              style: {
-                position: arg(absolute, 'absolute'),
-                top: arg(top, '0'),
-                zIndex: arg(zIndex, '10', '0'),
-                background: 'white',
-              },
-            },
-            m(
-              '',
-              `A very simple portal - a div rendered outside of the normal
-              flow of the page`,
-            ),
-          ),
-      ];
-    },
-  };
-}
-
-function lorem() {
-  const text = `Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-      tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
-      veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
-      commodo consequat.Duis aute irure dolor in reprehenderit in voluptate
-      velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
-      cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id
-      est laborum.`;
-  return m('', {style: {width: '200px'}}, text);
-}
-
-function ControlledPopup() {
-  let popupOpen = false;
-
-  return {
-    view: function () {
-      return m(
-        Popup,
-        {
-          trigger: m(Button, {
-            label: `${popupOpen ? 'Close' : 'Open'} Popup`,
-            onclick: () => {
-              popupOpen = true;
-            },
-          }),
-          isOpen: popupOpen,
-          onChange: (shouldOpen: boolean) => (popupOpen = shouldOpen),
-        },
-        m(Button, {
-          label: 'Close Popup',
-          onclick: () => {
-            popupOpen = false;
-          },
-        }),
-      );
-    },
-  };
-}
-
-function recursiveTreeNode(): m.Children {
-  return m(LazyTreeNode, {
-    left: 'Recursive',
-    right: '...',
-    fetchData: async () => {
-      // await new Promise((r) => setTimeout(r, 1000));
-      return () => recursiveTreeNode();
-    },
-  });
-}
-
-interface File {
-  name: string;
-  size: string;
-  date: string;
-  children?: File[];
-}
-
-const files: File[] = [
-  {
-    name: 'foo',
-    size: '10MB',
-    date: '2023-04-02',
-  },
-  {
-    name: 'bar',
-    size: '123KB',
-    date: '2023-04-08',
-    children: [
-      {
-        name: 'baz',
-        size: '4KB',
-        date: '2023-05-07',
-      },
-      {
-        name: 'qux',
-        size: '18KB',
-        date: '2023-05-28',
-        children: [
-          {
-            name: 'quux',
-            size: '4KB',
-            date: '2023-05-07',
-          },
-          {
-            name: 'corge',
-            size: '18KB',
-            date: '2023-05-28',
-            children: [
-              {
-                name: 'grault',
-                size: '4KB',
-                date: '2023-05-07',
-              },
-              {
-                name: 'garply',
-                size: '18KB',
-                date: '2023-05-28',
-              },
-              {
-                name: 'waldo',
-                size: '87KB',
-                date: '2023-05-02',
-              },
-            ],
-          },
-        ],
-      },
-    ],
-  },
-  {
-    name: 'fred',
-    size: '8KB',
-    date: '2022-12-27',
-  },
+const WIDGET_SECTIONS: WidgetSection[] = [
+  {id: 'anchor', label: 'Anchor', view: anchor},
+  {id: 'button', label: 'Button', view: renderButtonDemo},
+  {id: 'button-group', label: 'ButtonGroup', view: renderButtonGroupDemo},
+  {id: 'callout', label: 'Callout', view: renderCallout},
+  {id: 'card-stack', label: 'CardStack', view: cardStack},
+  {id: 'card', label: 'Card', view: renderCard},
+  {id: 'checkbox', label: 'Checkbox', view: renderCheckbox},
+  {id: 'chip', label: 'Chip', view: renderChip},
+  {id: 'codesnippet', label: 'CodeSnippet', view: renderCodeSnippet},
+  {id: 'copyablelink', label: 'CopyableLink', view: renderCopyableLink},
+  {id: 'cursor-tooltip', label: 'CursorTooltip', view: cursorTooltip},
+  {id: 'datagrid', label: 'DataGrid', view: renderDataGrid},
+  {id: 'editor', label: 'Editor', view: renderEditor},
+  {id: 'emptystate', label: 'EmptyState', view: renderEmptyState},
+  {id: 'form', label: 'Form', view: renderForm},
+  {id: 'grid', label: 'Grid', view: renderGrid},
+  {id: 'hotkey', label: 'Hotkey', view: renderHotkey},
+  {id: 'icon', label: 'Icon', view: renderIcon},
+  {id: 'menu', label: 'Menu', view: renderMenu},
+  {id: 'middleellipsis', label: 'MiddleEllipsis', view: renderMiddleEllipsis},
+  {id: 'modal', label: 'Modal', view: renderModal},
+  {id: 'multiselect', label: 'Multiselect', view: renderMultiselect},
+  {id: 'nodegraph', label: 'NodeGraph', view: renderNodeGraph},
+  {id: 'popup', label: 'Popup', view: renderPopup},
+  {id: 'popup-menu', label: 'PopupMenu', view: popupMenuDemo},
+  {id: 'portal', label: 'Portal', view: renderPortal},
+  {id: 'segmented-buttons', label: 'SegmentedButtons', view: segmentedButtons},
+  {id: 'select', label: 'Select', view: renderSelect},
+  {id: 'spinner', label: 'Spinner', view: renderSpinner},
+  {id: 'split-panel', label: 'Split Panel', view: renderSplitPanel},
+  {id: 'switch', label: 'Switch', view: renderSwitch},
+  {id: 'tabstrip', label: 'TabStrip', view: renderTabStrip},
+  {id: 'taginput', label: 'TagInput', view: renderTagInput},
+  {id: 'textinput', label: 'TextInput', view: renderTextInput},
+  {id: 'textparagraph', label: 'TextParagraph', view: renderTextParagraph},
+  {id: 'tooltip', label: 'Tooltip', view: renderTooltip},
+  {id: 'trackshell', label: 'TrackShell', view: renderTrackShell},
+  {id: 'tree', label: 'Tree', view: renderTree},
+  {id: 'treetable', label: 'TreeTable', view: renderTreeTable},
+  {id: 'vegaview', label: 'VegaView', view: renderVegaView},
+  {id: 'virtualcanvas', label: 'VirtualCanvas', view: renderVirtualCanvas},
 ];
 
-function TagInputDemo() {
-  const tags: string[] = ['foo', 'bar', 'baz'];
-  let tagInputValue: string = '';
-
-  return {
-    view: () => {
-      return m(TagInput, {
-        tags,
-        value: tagInputValue,
-        onTagAdd: (tag) => {
-          tags.push(tag);
-          tagInputValue = '';
-        },
-        onChange: (value) => {
-          tagInputValue = value;
-        },
-        onTagRemove: (index) => {
-          tags.splice(index, 1);
-        },
-      });
-    },
-  };
+export interface WidgetsPageAttrs {
+  readonly app: App;
+  readonly subpage?: string;
 }
 
-function SegmentedButtonsDemo({attrs}: {attrs: {}}) {
-  let selectedIdx = 0;
-  return {
-    view: () => {
-      return m(SegmentedButtons, {
-        ...attrs,
-        options: [{label: 'Yes'}, {label: 'Maybe'}, {label: 'No'}],
-        selectedOption: selectedIdx,
-        onOptionSelected: (num) => {
-          selectedIdx = num;
-        },
-      });
-    },
-  };
-}
+export class WidgetsPage implements m.ClassComponent<WidgetsPageAttrs> {
+  view({attrs}: m.Vnode<WidgetsPageAttrs>) {
+    const currentSection = WIDGET_SECTIONS.find(
+      (s) => `/${s.id}` === attrs.subpage,
+    );
 
-function RadioButtonGroupDemo() {
-  let setting: 'yes' | 'maybe' | 'no' = 'no';
-  console.log(setting);
-  return {
-    view: ({attrs}: m.Vnode<ButtonAttrs>) => {
-      return m(ButtonGroup, [
-        m(Button, {
-          ...attrs,
-          label: 'Yes',
-          active: setting === 'yes',
-          onclick: () => {
-            setting = 'yes';
-          },
-        }),
-        m(Button, {
-          ...attrs,
-          label: 'Maybe',
-          active: setting === 'maybe',
-          onclick: () => {
-            setting = 'maybe';
-          },
-        }),
-        m(Button, {
-          ...attrs,
-          label: 'No',
-          active: setting === 'no',
-          onclick: () => {
-            setting = 'no';
-          },
-        }),
-      ]);
-    },
-  };
-}
-
-export class WidgetsPage implements m.ClassComponent<{app: App}> {
-  view({attrs}: m.Vnode<{app: App}>) {
     return m(
       '.pf-widgets-page',
-      m('h1', 'Widgets'),
-      m(WidgetShowcase, {
-        label: 'Button',
-        renderWidget: ({
-          label,
-          icon,
-          rightIcon,
-          showAsGrid,
-          showInlineWithText,
-          ...rest
-        }) =>
-          Boolean(showAsGrid)
-            ? m(
-                '',
-                {
-                  style: {
-                    display: 'grid',
-                    gridTemplateColumns: 'auto auto auto',
-                    gap: '4px',
-                  },
+      // Left sidebar menu
+      m(
+        '.pf-widgets-page__menu',
+        {key: 'widgets-menu'},
+        m(
+          'nav',
+          WIDGET_SECTIONS.map((sec) =>
+            m(
+              '.pf-widgets-page__menu-item',
+              {
+                className: classNames(sec === currentSection && 'pf-active'),
+                onclick: () => {
+                  window.location.hash = `#!/widgets/${sec.id}`;
                 },
-                Object.values(Intent).map((intent) => {
-                  return Object.values(ButtonVariant).map((variant) => {
-                    return m(Button, {
-                      style: {
-                        width: '80px',
-                      },
-                      ...rest,
-                      label: variant,
-                      variant,
-                      intent,
-                    });
-                  });
-                }),
-              )
-            : m('', [
-                Boolean(showInlineWithText) && 'Inline',
-                m(Button, {
-                  icon: arg(icon, 'send'),
-                  rightIcon: arg(rightIcon, 'arrow_forward'),
-                  label: arg(label, 'Button', ''),
-                  onclick: () => console.log('button pressed'),
-                  ...rest,
-                }),
-                Boolean(showInlineWithText) && 'text',
-              ]),
-        initialOpts: {
-          label: true,
-          icon: true,
-          rightIcon: false,
-          disabled: false,
-          intent: new EnumOption(Intent.None, Object.values(Intent)),
-          active: false,
-          compact: false,
-          loading: false,
-          variant: new EnumOption(
-            ButtonVariant.Filled,
-            Object.values(ButtonVariant),
-          ),
-          showAsGrid: false,
-          showInlineWithText: false,
-          rounded: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Segmented Buttons',
-        description: `
-          Segmented buttons are a group of buttons where one of them is
-          'selected'; they act similar to a set of radio buttons.
-        `,
-        renderWidget: (opts) => m(SegmentedButtonsDemo, opts),
-        initialOpts: {
-          disabled: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'ButtonGroup',
-        renderWidget: (opts) =>
-          m(Stack, [
-            m(ButtonGroup, [
-              m(Button, {
-                label: 'Commit',
-                ...opts,
-              }),
-              m(Button, {
-                icon: Icons.ContextMenu,
-                ...opts,
-              }),
-            ]),
-            m(RadioButtonGroupDemo, opts),
-          ]),
-        initialOpts: {
-          variant: new EnumOption(
-            ButtonVariant.Filled,
-            Object.values(ButtonVariant),
-          ),
-          disabled: false,
-          intent: new EnumOption(Intent.None, Object.values(Intent)),
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Checkbox',
-        renderWidget: (opts) => m(Checkbox, {label: 'Checkbox', ...opts}),
-        initialOpts: {
-          disabled: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Switch',
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        renderWidget: ({label, labelLeft, ...rest}: any) =>
-          m(Switch, {
-            label: arg(label, 'Switch'),
-            labelLeft: arg(labelLeft, 'Left Label'),
-            ...rest,
-          }),
-        initialOpts: {
-          label: true,
-          labelLeft: false,
-          disabled: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Anchor',
-        renderWidget: ({icon, showInlineWithText, long}) =>
-          m('', [
-            Boolean(showInlineWithText) && 'Inline',
-            m(
-              Anchor,
-              {
-                icon: arg(icon, 'open_in_new'),
-                href: 'https://perfetto.dev/docs/',
-                target: '_blank',
               },
-              Boolean(long)
-                ? 'This is some really long text and it will probably overflow the container'
-                : 'Link',
+              sec.label,
             ),
-            Boolean(showInlineWithText) && 'text',
-          ]),
-
-        initialOpts: {
-          icon: true,
-          showInlineWithText: false,
-          long: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Text Input',
-        renderWidget: ({placeholder, leftIcon, ...rest}) =>
-          m(TextInput, {
-            placeholder: arg(placeholder, 'Placeholder...', ''),
-            leftIcon: arg(leftIcon, 'search'),
-            ...rest,
-          }),
-        initialOpts: {
-          placeholder: true,
-          disabled: false,
-          leftIcon: true,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Select',
-        renderWidget: (opts) =>
-          m(Select, opts, [
-            m('option', {value: 'foo', label: 'Foo'}),
-            m('option', {value: 'bar', label: 'Bar'}),
-            m('option', {value: 'baz', label: 'Baz'}),
-          ]),
-        initialOpts: {
-          disabled: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Empty State',
-        renderWidget: ({header, content}) =>
-          m(
-            EmptyState,
-            {
-              title: arg(header, 'No search results found...'),
-            },
-            arg(content, m(Button, {label: 'Try again'})),
           ),
-        initialOpts: {
-          header: true,
-          content: true,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Card',
-        description: `A card is a simple container with a shadow and rounded
-          corners. It can be used to display grouped content in a visually
-          appealing way.`,
-        renderWidget: ({interactive}) =>
-          m(Card, {interactive}, [
-            m('h1', {style: {margin: 'unset'}}, 'Welcome!'),
-            m('p', 'Would you like to start your journey?'),
-            m(Stack, {orientation: 'horizontal'}, [
-              m(Button, {
-                variant: ButtonVariant.Filled,
-                label: 'No thanks...',
-              }),
-              m(Button, {
-                intent: Intent.Primary,
-                variant: ButtonVariant.Filled,
-                label: "Let's go!",
-              }),
-            ]),
-          ]),
-        initialOpts: {interactive: true},
-      }),
-      m(WidgetShowcase, {
-        label: 'CardStack',
-        description: `A container component that can be used to display
-          multiple Card elements in a vertical stack. Cards placed in this list
-          automatically have their borders adjusted to appear as one continuous
-          card with thin borders between them.`,
-        renderWidget: ({direction, interactive}) =>
-          m(CardStack, {direction}, [
-            m(Card, {interactive}, m(Switch, {label: 'Option 1'})),
-            m(Card, {interactive}, m(Switch, {label: 'Option 2'})),
-            m(Card, {interactive}, m(Switch, {label: 'Option 3'})),
-          ]),
-        initialOpts: {
-          direction: new EnumOption('vertical', ['vertical', 'horizontal']),
-          interactive: true,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'CopyableLink',
-        renderWidget: ({noicon}) =>
-          m(CopyableLink, {
-            noicon: arg(noicon, true),
-            url: 'https://perfetto.dev/docs/',
-          }),
-        initialOpts: {
-          noicon: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'CopyToClipboardButton',
-        renderWidget: (opts) =>
-          m(CopyToClipboardButton, {
-            textToCopy: 'Text to copy',
-            ...opts,
-          }),
-        initialOpts: {
-          label: 'Copy',
-          variant: new EnumOption(
-            ButtonVariant.Outlined,
-            Object.values(ButtonVariant),
-          ),
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Portal',
-        description: `A portal is a div rendered out of normal flow
-          of the hierarchy.`,
-        renderWidget: (opts) => m(PortalButton, opts),
-        initialOpts: {
-          absolute: true,
-          zIndex: true,
-          top: true,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Popup',
-        description: `A popup is a nicely styled portal element whose position is
-        dynamically updated to appear to float alongside a specific element on
-        the page, even as the element is moved and scrolled around.`,
-        renderWidget: (opts) =>
-          m(
-            Popup,
-            {
-              trigger: m(Button, {label: 'Toggle Popup'}),
-              ...opts,
-            },
-            lorem(),
-          ),
-        initialOpts: {
-          position: new EnumOption(
-            PopupPosition.Auto,
-            Object.values(PopupPosition),
-          ),
-          closeOnEscape: true,
-          closeOnOutsideClick: true,
-          isContextMenu: false,
-          positionAtCursor: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Controlled Popup',
-        description: `The open/close state of a controlled popup is passed in via
-        the 'isOpen' attribute. This means we can get open or close the popup
-        from wherever we like. E.g. from a button inside the popup.
-        Keeping this state external also means we can modify other parts of the
-        page depending on whether the popup is open or not, such as the text
-        on this button.
-        Note, this is the same component as the popup above, but used in
-        controlled mode.`,
-        renderWidget: (opts) => m(ControlledPopup, opts),
-        initialOpts: {},
-      }),
-      m(WidgetShowcase, {
-        label: 'Icon',
-        renderWidget: (opts) => m(Icon, {icon: 'star', ...opts}),
-        initialOpts: {
-          filled: false,
-          intent: new EnumOption(Intent.None, Object.values(Intent)),
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Tooltip',
-        description: `A tooltip is a hover-only, useful as an alternative to the browser's inbuilt 'title' tooltip.`,
-        renderWidget: (opts) =>
-          m(
-            Tooltip,
-            {
-              trigger: m(Icon, {icon: 'Warning'}),
-              ...opts,
-            },
-            lorem(),
-          ),
-        initialOpts: {
-          position: new EnumOption(
-            PopupPosition.Auto,
-            Object.values(PopupPosition),
-          ),
-          showArrow: true,
-          offset: 0,
-          edgeOffset: 0,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'MultiSelect panel',
-        renderWidget: ({...rest}) =>
-          m(MultiSelect, {
-            options: Object.entries(options).map(([key, value]) => {
-              return {
-                id: key,
-                name: key,
-                checked: value,
-              };
-            }),
-            onChange: (diffs: MultiSelectDiff[]) => {
-              diffs.forEach(({id, checked}) => {
-                options[id] = checked;
-              });
-            },
-            ...rest,
-          }),
-        initialOpts: {
-          repeatCheckedItemsAtTop: false,
-          fixedSize: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Popup with MultiSelect',
-        renderWidget: ({icon, ...rest}) =>
-          m(PopupMultiSelect, {
-            options: Object.entries(options).map(([key, value]) => {
-              return {
-                id: key,
-                name: key,
-                checked: value,
-              };
-            }),
-            position: PopupPosition.Top,
-            label: 'Multi Select',
-            icon: arg(icon, Icons.LibraryAddCheck),
-            onChange: (diffs: MultiSelectDiff[]) => {
-              diffs.forEach(({id, checked}) => {
-                options[id] = checked;
-              });
-            },
-            ...rest,
-          }),
-        initialOpts: {
-          icon: true,
-          showNumSelected: true,
-          repeatCheckedItemsAtTop: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'MultiselectInput',
-        description: `Tag input with options`,
-        renderWidget: () => {
-          return m(MultiselectInputDemo);
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Menu',
-        renderWidget: () =>
-          m(
-            Menu,
-            m(MenuItem, {label: 'New', icon: 'add'}),
-            m(MenuItem, {label: 'Open', icon: 'folder_open'}),
-            m(MenuItem, {label: 'Save', icon: 'save', disabled: true}),
-            m(MenuDivider),
-            m(MenuItem, {label: 'Delete', icon: 'delete'}),
-            m(MenuDivider),
-            m(
-              MenuItem,
-              {label: 'Share', icon: 'share'},
-              m(MenuItem, {label: 'Everyone', icon: 'public'}),
-              m(MenuItem, {label: 'Friends', icon: 'group'}),
-              m(
-                MenuItem,
-                {label: 'Specific people', icon: 'person_add'},
-                m(MenuItem, {label: 'Alice', icon: 'person'}),
-                m(MenuItem, {label: 'Bob', icon: 'person'}),
+        ),
+      ),
+      // Main content area
+      m(
+        '.pf-widgets-page__content-container',
+        {key: currentSection ? currentSection.id : 'no-section'},
+        m(
+          '.pf-widgets-page__content',
+          currentSection
+            ? currentSection.view(attrs.app)
+            : m(
+                '.pf-widget-intro',
+                m('h1', 'Widgets Showcase'),
+                m('p', [
+                  'This showcase demonstrates the reusable UI components available in the Perfetto UI framework.',
+                ]),
+                m('p', [
+                  'Click on a widget name in the menu on the left to view interactive demos and examples.',
+                ]),
               ),
-            ),
-            m(
-              MenuItem,
-              {label: 'More', icon: 'more_horiz'},
-              m(MenuItem, {label: 'Query', icon: 'database'}),
-              m(MenuItem, {label: 'Download', icon: 'download'}),
-              m(MenuItem, {label: 'Clone', icon: 'copy_all'}),
-            ),
-          ),
-      }),
-      m(WidgetShowcase, {
-        label: 'PopupMenu',
-        renderWidget: (opts) =>
-          m(
-            PopupMenu,
-            {
-              trigger: m(Button, {
-                label: 'Menu',
-                rightIcon: Icons.ContextMenu,
-              }),
-              ...opts,
-            },
-            m(MenuItem, {label: 'New', icon: 'add'}),
-            m(MenuItem, {label: 'Open', icon: 'folder_open'}),
-            m(MenuItem, {label: 'Save', icon: 'save', disabled: true}),
-            m(MenuDivider),
-            m(MenuItem, {label: 'Delete', icon: 'delete'}),
-            m(MenuDivider),
-            m(
-              MenuItem,
-              {label: 'Share', icon: 'share'},
-              m(MenuItem, {label: 'Everyone', icon: 'public'}),
-              m(MenuItem, {label: 'Friends', icon: 'group'}),
-              m(
-                MenuItem,
-                {label: 'Specific people', icon: 'person_add'},
-                m(MenuItem, {label: 'Alice', icon: 'person'}),
-                m(MenuItem, {label: 'Bob', icon: 'person'}),
-              ),
-            ),
-            m(
-              MenuItem,
-              {label: 'More', icon: 'more_horiz'},
-              m(MenuItem, {label: 'Query', icon: 'database'}),
-              m(MenuItem, {label: 'Download', icon: 'download'}),
-              m(MenuItem, {label: 'Clone', icon: 'copy_all'}),
-            ),
-          ),
-        initialOpts: {
-          position: new EnumOption(
-            PopupPosition.Bottom,
-            Object.values(PopupPosition),
-          ),
-          isContextMenu: false,
-          positionAtCursor: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'CursorTooltip',
-        description: 'A tooltip that follows the mouse around.',
-        renderWidget: () => m(CursorTooltipShowcase),
-      }),
-      m(WidgetShowcase, {
-        label: 'Spinner',
-        description: `Simple spinner, rotates forever.
-            Width and height match the font size.`,
-        renderWidget: ({fontSize, easing}) =>
-          m('', {style: {fontSize}}, m(Spinner, {easing})),
-        initialOpts: {
-          fontSize: new EnumOption('16px', [
-            '12px',
-            '16px',
-            '24px',
-            '32px',
-            '64px',
-            '128px',
-          ]),
-          easing: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Tree',
-        description: `Hierarchical tree with left and right values aligned to
-        a grid.`,
-        renderWidget: (opts) =>
-          m(
-            Tree,
-            opts,
-            m(TreeNode, {left: 'Name', right: 'my_event', icon: 'badge'}),
-            m(TreeNode, {left: 'CPU', right: '2', icon: 'memory'}),
-            m(TreeNode, {
-              left: 'Start time',
-              right: '1s 435ms',
-              icon: 'schedule',
-            }),
-            m(TreeNode, {left: 'Duration', right: '86ms', icon: 'timer'}),
-            m(TreeNode, {
-              left: 'SQL',
-              right: m(
-                PopupMenu,
-                {
-                  position: PopupPosition.RightStart,
-                  trigger: m(
-                    Anchor,
-                    {
-                      icon: Icons.ContextMenu,
-                    },
-                    'SELECT * FROM ftrace_event WHERE id = 123',
-                  ),
-                },
-                m(MenuItem, {
-                  label: 'Copy SQL Query',
-                  icon: 'content_copy',
-                }),
-                m(MenuItem, {
-                  label: 'Execute Query in new tab',
-                  icon: 'open_in_new',
-                }),
-              ),
-            }),
-            m(TreeNode, {
-              icon: 'account_tree',
-              left: 'Process',
-              right: m(Anchor, {icon: 'open_in_new'}, '/bin/foo[789]'),
-            }),
-            m(TreeNode, {
-              left: 'Thread',
-              right: m(Anchor, {icon: 'open_in_new'}, 'my_thread[456]'),
-            }),
-            m(
-              TreeNode,
-              {
-                left: 'Args',
-                summary: 'foo: string, baz: string, quux: string[4]',
-              },
-              m(TreeNode, {left: 'foo', right: 'bar'}),
-              m(TreeNode, {left: 'baz', right: 'qux'}),
-              m(
-                TreeNode,
-                {left: 'quux', summary: 'string[4]'},
-                m(TreeNode, {left: '[0]', right: 'corge'}),
-                m(TreeNode, {left: '[1]', right: 'grault'}),
-                m(TreeNode, {left: '[2]', right: 'garply'}),
-                m(TreeNode, {left: '[3]', right: 'waldo'}),
-              ),
-            ),
-            m(LazyTreeNode, {
-              left: 'Lazy',
-              icon: 'bedtime',
-              fetchData: async () => {
-                await new Promise((r) => setTimeout(r, 1000));
-                return () => m(TreeNode, {left: 'foo'});
-              },
-            }),
-            m(LazyTreeNode, {
-              left: 'Dynamic',
-              unloadOnCollapse: true,
-              icon: 'bedtime',
-              fetchData: async () => {
-                await new Promise((r) => setTimeout(r, 1000));
-                return () => m(TreeNode, {left: 'foo'});
-              },
-            }),
-            recursiveTreeNode(),
-          ),
-        wide: true,
-      }),
-      m(WidgetShowcase, {
-        label: 'Form',
-        renderWidget: () => renderForm('form'),
-      }),
-      m(WidgetShowcase, {
-        label: 'Nested Popups',
-        renderWidget: () =>
-          m(
-            Popup,
-            {
-              trigger: m(Button, {label: 'Open the popup'}),
-            },
-            m(ButtonBar, [
-              m(
-                PopupMenu,
-                {
-                  trigger: m(Button, {label: 'Select an option'}),
-                },
-                m(MenuItem, {label: 'Option 1'}),
-                m(MenuItem, {label: 'Option 2'}),
-              ),
-              m(Button, {
-                label: 'Done',
-                dismissPopup: true,
-              }),
-            ]),
-          ),
-      }),
-      m(WidgetShowcase, {
-        label: 'Callout',
-        renderWidget: ({icon, ...opts}) =>
-          m(
-            Callout,
-            {
-              icon: Boolean(icon) ? 'info' : undefined,
-              ...opts,
-            },
-            'Lorem ipsum dolor sit amet, consectetur adipiscing elit. ' +
-              'Nulla rhoncus tempor neque, sed malesuada eros dapibus vel. ' +
-              'Aliquam in ligula vitae tortor porttitor laoreet iaculis ' +
-              'finibus est.',
-          ),
-        initialOpts: {
-          intent: new EnumOption(Intent.None, Object.values(Intent)),
-          dismissable: false,
-          icon: true,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Editor',
-        renderWidget: () =>
-          m(Editor, {
-            language: 'perfetto-sql',
-            onUpdate: (text) => {
-              parseAndPrintTree(text);
-            },
-          }),
-      }),
-      m(WidgetShowcase, {
-        label: 'VegaView',
-        renderWidget: (opt) =>
-          m(VegaView, {
-            spec: getExampleSpec(opt.exampleSpec),
-            data: getExampleData(opt.exampleData),
-          }),
-        initialOpts: {
-          exampleSpec: new EnumOption(
-            SpecExample.BarChart,
-            Object.values(SpecExample),
-          ),
-          exampleData: new EnumOption(
-            DataExample.English,
-            Object.values(DataExample),
-          ),
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Form within PopupMenu',
-        description: `A form placed inside a popup menu works just fine,
-              and the cancel/submit buttons also dismiss the popup. A bit more
-              margin is added around it too, which improves the look and feel.`,
-        renderWidget: () =>
-          m(
-            PopupMenu,
-            {
-              trigger: m(Button, {label: 'Popup!'}),
-            },
-            m(
-              MenuItem,
-              {
-                label: 'Open form...',
-              },
-              renderForm('popup-form'),
-            ),
-          ),
-      }),
-      m(WidgetShowcase, {
-        label: 'Hotkey',
-        renderWidget: (opts) => {
-          if (opts.platform === 'auto') {
-            return m(HotkeyGlyphs, {hotkey: opts.hotkey as Hotkey});
-          } else {
-            const platform = opts.platform as Platform;
-            return m(HotkeyGlyphs, {
-              hotkey: opts.hotkey as Hotkey,
-              spoof: platform,
-            });
-          }
-        },
-        initialOpts: {
-          hotkey: 'Mod+Shift+P',
-          platform: new EnumOption('auto', ['auto', 'Mac', 'PC']),
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Text Paragraph',
-        description: `A basic formatted text paragraph with wrapping. If
-              it is desirable to preserve the original text format/line breaks,
-              set the compressSpace attribute to false.`,
-        renderWidget: (opts) => {
-          return m(TextParagraph, {
-            text: `Lorem ipsum dolor sit amet, consectetur adipiscing
-                         elit. Nulla rhoncus tempor neque, sed malesuada eros
-                         dapibus vel. Aliquam in ligula vitae tortor porttitor
-                         laoreet iaculis finibus est.`,
-            compressSpace: opts.compressSpace,
-          });
-        },
-        initialOpts: {
-          compressSpace: true,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Multi Paragraph Text',
-        description: `A wrapper for multiple paragraph widgets.`,
-        renderWidget: () => {
-          return m(
-            MultiParagraphText,
-            m(TextParagraph, {
-              text: `Lorem ipsum dolor sit amet, consectetur adipiscing
-                         elit. Nulla rhoncus tempor neque, sed malesuada eros
-                         dapibus vel. Aliquam in ligula vitae tortor porttitor
-                         laoreet iaculis finibus est.`,
-              compressSpace: true,
-            }),
-            m(TextParagraph, {
-              text: `Sed ut perspiciatis unde omnis iste natus error sit
-                         voluptatem accusantium doloremque laudantium, totam rem
-                         aperiam, eaque ipsa quae ab illo inventore veritatis et
-                         quasi architecto beatae vitae dicta sunt explicabo.
-                         Nemo enim ipsam voluptatem quia voluptas sit aspernatur
-                         aut odit aut fugit, sed quia consequuntur magni dolores
-                         eos qui ratione voluptatem sequi nesciunt.`,
-              compressSpace: true,
-            }),
-          );
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Modal',
-        description: `Shows a dialog box in the center of the screen over the
-                      top of other elements.`,
-        renderWidget: () =>
-          m(Button, {
-            label: 'Show Modal',
-            onclick: () => {
-              showModal({
-                title: 'Attention',
-                icon: Icons.Help,
-                content: () => [
-                  m('', 'This is a modal dialog'),
-                  m(
-                    Popup,
-                    {
-                      trigger: m(Button, {
-                        variant: ButtonVariant.Filled,
-                        label: 'Open Popup',
-                      }),
-                    },
-                    'Popup content',
-                  ),
-                ],
-                buttons: [
-                  {
-                    text: 'Cancel',
-                  },
-                  {
-                    text: 'OK',
-                    primary: true,
-                  },
-                ],
-              });
-            },
-          }),
-      }),
-      m(WidgetShowcase, {
-        label: 'Advanced Modal',
-        description: `A helper for modal dialog.`,
-        renderWidget: () => m(ModalShowcase),
-      }),
-      m(WidgetShowcase, {
-        label: 'TreeTable',
-        description: `Hierarchical tree with multiple columns`,
-        renderWidget: () => {
-          const attrs: TreeTableAttrs<File> = {
-            rows: files,
-            getChildren: (file) => file.children,
-            columns: [
-              {name: 'Name', getData: (file) => file.name},
-              {name: 'Size', getData: (file) => file.size},
-              {name: 'Date', getData: (file) => file.date},
-            ],
-          };
-          return m(TreeTable<File>, attrs);
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Tag Input',
-        description: `
-          TagInput displays Tag elements inside an input, followed by an
-          interactive text input. The container is styled to look like a
-          TextInput, but the actual editable element appears after the last tag.
-          Clicking anywhere on the container will focus the text input.`,
-        renderWidget: () => m(TagInputDemo),
-      }),
-      m(WidgetShowcase, {
-        label: 'Middle Ellipsis',
-        description: `
-          Sometimes the start and end of a bit of text are more important than
-          the middle. This element puts the ellipsis in the midde if the content
-          is too wide for its container.`,
-        renderWidget: (opts) =>
-          m(
-            'div',
-            {style: {width: Boolean(opts.squeeze) ? '150px' : '450px'}},
-            m(MiddleEllipsis, {
-              text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit',
-            }),
-          ),
-        initialOpts: {
-          squeeze: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Chip',
-        description: `A little chip or tag`,
-        renderWidget: (opts) => {
-          const {icon, ...rest} = opts;
-          return m(
-            Stack,
-            {orientation: 'horizontal'},
-            m(Chip, {
-              label: 'Foo',
-              icon: icon === true ? 'info' : undefined,
-              ...rest,
-            }),
-            m(Chip, {
-              label: 'Bar',
-              icon: icon === true ? 'warning' : undefined,
-              ...rest,
-            }),
-            m(Chip, {
-              label: 'Baz',
-              icon: icon === true ? 'error' : undefined,
-              ...rest,
-            }),
-          );
-        },
-        initialOpts: {
-          intent: new EnumOption(Intent.None, Object.values(Intent)),
-          icon: true,
-          compact: false,
-          rounded: false,
-          disabled: false,
-          removable: true,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'TrackShell',
-        description: `The Mithril parts of a track (the shell, mainly).`,
-        renderWidget: (opts) => {
-          const {buttons, chips, multipleTracks, error, ...rest} = opts;
-          const dummyButtons = () => [
-            m(Button, {icon: 'info', compact: true}),
-            m(Button, {icon: 'settings', compact: true}),
-          ];
-          const dummyChips = () => ['foo', 'bar'];
-
-          const renderTrack = (children?: m.Children) =>
-            m(
-              TrackShell,
-              {
-                buttons: Boolean(buttons) ? dummyButtons() : undefined,
-                chips: Boolean(chips) ? dummyChips() : undefined,
-                error: Boolean(error)
-                  ? new Error('An error has occurred')
-                  : undefined,
-                ...rest,
-              },
-              children,
-            );
-
-          return m(
-            '',
-            {
-              style: {width: '500px', boxShadow: '0px 0px 1px 1px lightgray'},
-            },
-            Boolean(multipleTracks)
-              ? [renderTrack(), renderTrack(), renderTrack()]
-              : renderTrack(),
-          );
-        },
-        initialOpts: {
-          title: 'This is the title of the track',
-          subtitle: 'This is the subtitle of the track',
-          buttons: true,
-          chips: true,
-          heightPx: 32,
-          collapsible: true,
-          collapsed: true,
-          summary: false,
-          highlight: false,
-          error: false,
-          multipleTracks: false,
-          reorderable: false,
-          depth: 0,
-          lite: false,
-        },
-      }),
-      m(WidgetShowcase, {
-        label: 'Virtual Overlay Canvas',
-        description: `A scrolling container that draws a virtual canvas over
-          the top of it's content and keeps it in the viewport to make it appear
-          like there is one big canvas over the top of the content.`,
-        renderWidget: () => {
-          const width = 200;
-          const rowCount = 65536;
-          const rowHeight = 20;
-          return m(
-            VirtualOverlayCanvas,
-            {
-              className: 'pf-virtual-canvas',
-              overflowY: 'auto',
-              onCanvasRedraw({ctx, canvasRect}) {
-                ctx.strokeStyle = 'red';
-                ctx.lineWidth = 1;
-
-                ctx.font = '20px Arial';
-                ctx.fillStyle = 'black';
-                ctx.textAlign = 'left';
-                ctx.textBaseline = 'top';
-
-                for (let i = 0; i < rowCount; i++) {
-                  const rect = {
-                    left: 0,
-                    top: i * rowHeight,
-                    right: width,
-                    bottom: i * rowHeight + rowHeight,
-                  };
-                  if (canvasRect.overlaps(rect)) {
-                    ctx.strokeRect(0, i * rowHeight, width, rowHeight);
-                    ctx.fillText(`Row: ${i}`, 0, i * rowHeight);
-                  }
-                }
-              },
-            },
-            m('', {
-              style: {height: `${rowCount * rowHeight}px`, width: `${width}px`},
-            }),
-          );
-        },
-        initialOpts: {},
-      }),
-
-      m(WidgetShowcase, {
-        label: 'SplitPanel',
-        description: `Resizeable split panel with optional tabs.`,
-        renderWidget: (opts) => {
-          return m(
-            '',
-            {
-              style: {
-                height: '400px',
-                width: '400px',
-                border: 'solid 2px gray',
-              },
-            },
-            m(
-              SplitPanel,
-              {
-                leftHandleContent: [
-                  Boolean(opts.leftContent) && m(Button, {icon: 'Menu'}),
-                ],
-                drawerContent: 'Drawer Content',
-                tabs:
-                  Boolean(opts.tabs) &&
-                  m(
-                    '.pf-split-panel__tabs',
-                    m(
-                      Tab,
-                      {active: true, hasCloseButton: opts.showCloseButtons},
-                      'Foo',
-                    ),
-                    m(Tab, {hasCloseButton: opts.showCloseButtons}, 'Bar'),
-                  ),
-              },
-              'Main Content',
-            ),
-          );
-        },
-        initialOpts: {
-          leftContent: true,
-          tabs: true,
-          showCloseButtons: true,
-        },
-      }),
-
-      // Grid
-      renderGridDemo(),
-
-      // DataGrid
-      renderDataGridDemo(),
-
-      renderWidgetShowcase({
-        label: 'DataGrid (query backed)',
-        description: `An interactive data explorer and viewer - fetched from SQL.`,
-        renderWidget: ({
-          readonlyFilters,
-          readonlySorting,
-          aggregation,
-          ...rest
-        }) => {
-          const trace = attrs.app.trace;
-          if (trace) {
-            return m(
-              '',
-              {style: {height: '400px', width: '800px', overflow: 'hidden'}},
-              m(QueryDataGrid, {
-                ...rest,
-                engine: trace.engine,
-                query: `
-                  SELECT
-                    ts.id as id,
-                    dur,
-                    state,
-                    thread.name as thread_name,
-                    dur,
-                    io_wait,
-                    ucpu
-                  FROM thread_state ts
-                  JOIN thread USING(utid)
-                `,
-                fillHeight: true,
-                filters: readonlyFilters ? [] : undefined,
-                sorting: readonlySorting ? {direction: 'UNSORTED'} : undefined,
-                columns: [
-                  {
-                    name: 'id',
-                    title: 'ID',
-                    aggregation: aggregation ? 'COUNT' : undefined,
-                  },
-                  {
-                    name: 'dur',
-                    title: 'Duration',
-                    aggregation: aggregation ? 'SUM' : undefined,
-                  },
-                  {name: 'state', title: 'State'},
-                  {name: 'thread_name', title: 'Thread'},
-                  {name: 'ucpu', title: 'CPU'},
-                  {name: 'io_wait', title: 'IO Wait'},
-                ],
-              }),
-            );
-          } else {
-            return 'Load a trace to start';
-          }
-        },
-        initialOpts: {
-          showFiltersInToolbar: true,
-          readonlyFilters: false,
-          readonlySorting: false,
-          aggregation: false,
-        },
-      }),
-
-      m(WidgetShowcase, {
-        label: 'TabStrip',
-        description: `A simple tab strip`,
-        renderWidget: () => {
-          return m(TabStrip, {
-            tabs: [
-              {key: 'foo', title: 'Foo'},
-              {key: 'bar', title: 'Bar'},
-              {key: 'baz', title: 'Baz'},
-            ],
-            currentTabKey: currentTab,
-            onTabChange: (key) => {
-              currentTab = key;
-            },
-          });
-        },
-        initialOpts: {},
-      }),
-
-      m(WidgetShowcase, {
-        label: 'CodeSnippet',
-        renderWidget: ({wide}) =>
-          m(CodeSnippet, {
-            language: 'SQL',
-            text: Boolean(wide)
-              ? 'SELECT a_very_long_column_name, another_super_long_column_name, yet_another_ridiculously_long_column_name FROM a_table_with_an_unnecessarily_long_name WHERE some_condition_is_true AND another_condition_is_also_true;'
-              : 'SELECT * FROM slice LIMIT 10;',
-          }),
-        initialOpts: {
-          wide: false,
-        },
-      }),
-
-      m(WidgetShowcase, {
-        label: 'NodeGraph',
-        description:
-          'Scroll left and right to pan / pinch to zoom. Drag nodes from output to input ports to connect nodes. Click connectors or drag outputs away from ports to remove connections.',
-        wide: true,
-        renderWidget: (opts) => m(NodeGraphDemo, opts),
-        initialOpts: {
-          multiselect: true,
-          titleBars: false,
-          accentBars: false,
-          allInputsLeft: false,
-          allOutputsRight: false,
-          colors: false,
-        },
-      }),
+        ),
+      ),
     );
   }
-}
-
-function CursorTooltipShowcase() {
-  let show = false;
-  return {
-    view() {
-      return m(
-        '',
-        {
-          style: {
-            width: '150px',
-            height: '150px',
-            border: '1px dashed gray',
-            userSelect: 'none',
-            color: 'gray',
-            textAlign: 'center',
-            lineHeight: '150px',
-          },
-          onmouseover: () => (show = true),
-          onmouseout: () => (show = false),
-        },
-        'Hover here...',
-        show && m(CursorTooltip, 'Hi!'),
-      );
-    },
-  };
-}
-
-function MultiselectInputDemo() {
-  const options = [
-    'foo',
-    'bar',
-    'baz',
-    'qux',
-    'quux',
-    'corge',
-    'grault',
-    'garply',
-    'waldo',
-    'fred',
-  ];
-  let selectedOptions: string[] = [];
-  return {
-    view() {
-      return m(MultiselectInput, {
-        options: options.map((o) => ({key: o, label: o})),
-        selectedOptions,
-        onOptionAdd: (key) => selectedOptions.push(key),
-        onOptionRemove: (key) => {
-          selectedOptions = selectedOptions.filter((x) => x !== key);
-        },
-      });
-    },
-  };
-}
-
-type QueryDataGridAttrs = Omit<DataGridAttrs, 'data'> & {
-  readonly query: string;
-  readonly engine: Engine;
-};
-
-function QueryDataGrid(vnode: m.Vnode<QueryDataGridAttrs>) {
-  const dataSource = new SQLDataSource(vnode.attrs.engine, vnode.attrs.query);
-
-  return {
-    view({attrs}: m.Vnode<QueryDataGridAttrs>) {
-      return m(DataGrid, {...attrs, data: dataSource});
-    },
-  };
-}
-
-class ModalShowcase implements m.ClassComponent {
-  private static counter = 0;
-
-  private static log(txt: string) {
-    const mwlogs = document.getElementById('mwlogs');
-    if (!mwlogs || !(mwlogs instanceof HTMLTextAreaElement)) return;
-    const time = new Date().toLocaleTimeString();
-    mwlogs.value += `[${time}] ${txt}\n`;
-    mwlogs.scrollTop = mwlogs.scrollHeight;
-  }
-
-  private static showModalDialog(staticContent = false) {
-    const id = `N=${++ModalShowcase.counter}`;
-    ModalShowcase.log(`Open ${id}`);
-    const logOnClose = () => ModalShowcase.log(`Close ${id}`);
-
-    let content;
-    if (staticContent) {
-      content = m(
-        '.pf-modal-pre',
-        'Content of the modal dialog.\nEnd of content',
-      );
-    } else {
-      // The humble counter is basically the VDOM 'Hello world'!
-      function CounterComponent() {
-        let counter = 0;
-        return {
-          view: () => {
-            return m(
-              '',
-              `Counter value: ${counter}`,
-              m(Button, {
-                label: 'Increment Counter',
-                onclick: () => ++counter,
-              }),
-            );
-          },
-        };
-      }
-      content = () => m(CounterComponent);
-    }
-    const closePromise = showModal({
-      title: `Modal dialog ${id}`,
-      buttons: [
-        {text: 'OK', action: () => ModalShowcase.log(`OK ${id}`)},
-        {text: 'Cancel', action: () => ModalShowcase.log(`Cancel ${id}`)},
-        {
-          text: 'Show another now',
-          action: () => ModalShowcase.showModalDialog(),
-        },
-        {
-          text: 'Show another in 2s',
-          action: () => setTimeout(() => ModalShowcase.showModalDialog(), 2000),
-        },
-      ],
-      content,
-    });
-    closePromise.then(logOnClose);
-  }
-
-  view() {
-    return m(
-      'div',
-      {
-        style: {
-          'display': 'flex',
-          'flex-direction': 'column',
-          'width': '100%',
-        },
-      },
-      m('textarea', {
-        id: 'mwlogs',
-        readonly: 'readonly',
-        rows: '8',
-        placeholder: 'Logs will appear here',
-      }),
-      m('input[type=button]', {
-        value: 'Show modal (static)',
-        onclick: () => ModalShowcase.showModalDialog(true),
-      }),
-      m('input[type=button]', {
-        value: 'Show modal (dynamic)',
-        onclick: () => ModalShowcase.showModalDialog(false),
-      }),
-    );
-  }
-} // class ModalShowcase
-
-function renderForm(id: string) {
-  return m(
-    Form,
-    {
-      submitLabel: 'Submit',
-      submitIcon: 'send',
-      cancelLabel: 'Cancel',
-      resetLabel: 'Reset',
-      onSubmit: () => window.alert('Form submitted!'),
-    },
-    m(FormLabel, {for: `${id}-foo`}, 'Foo'),
-    m(TextInput, {id: `${id}-foo`}),
-    m(FormLabel, {for: `${id}-bar`}, 'Bar'),
-    m(Select, {id: `${id}-bar`}, [
-      m('option', {value: 'foo', label: 'Foo'}),
-      m('option', {value: 'bar', label: 'Bar'}),
-      m('option', {value: 'baz', label: 'Baz'}),
-    ]),
-  );
 }
